// lib/client/types.gen.ts
export type Difficulty = 'лёгкий' | 'средний' | 'сложный';

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
    /**
     * Detail
     */
    detail?: Array<ValidationError>;
};

/**
 * TaskSchema
 */
export type TaskSchema = {
    id: string,
    /**
     * Subject
     */
    subject: string;
    theme: Theme;
    difficulty: Difficulty;
    /**
     * Title
     */
    title: string;
    /**
     * Task Text
     */
    task_text: string;
    /**
     * Hint
     */
    hint: string;
    /**
     * Answer
     */
    answer?: string | null;
    /**
     * Is Published
     */
    is_published?: boolean;
};

/**
 * Theme
 */
export type Theme = 'математика' | 'русский' | 'информатика' | 'физика';